feat: Add sign in page
This commit adds a new file `sign_in.dart` which contains the `SignIn` widget. This widget is responsible for rendering the sign in page in the authentication feature. The `SignIn` widget is a stateful widget that extends `StatefulWidget` and has a corresponding state class `_SignInState`. The `build` method of the `_SignInState` class returns a `Scaffold` widget. This commit is a feature addition and improves the organization of the codebase.