feat: Add sign up page
This commit adds a new file `sign_up.dart` which contains the `SignUp` widget. This widget is responsible for rendering the sign up page in the authentication feature. The `SignUp` widget is a stateful widget that extends `StatefulWidget` and has a corresponding state class `_SignUpState`. The `build` method of the `_SignUpState` class returns a `Placeholder` widget.