feat: Add Homescreen widget
This commit adds a new file `homescreen.dart` which contains the `Homescreen` widget. This widget is responsible for rendering the home screen in the app. The `Homescreen` widget is a stateful widget that extends `StatefulWidget` and has a corresponding state class `_HomescreenState`. The `build` method of the `_HomescreenState` class returns a `Placeholder` widget.