chore: Update dependencies and dev dependencies
This commit updates the dependencies and dev dependencies in the `pubspec.yaml` file. It removes the `flutter_bloc` package and adds the `flutter_riverpod` and `riverpod_annotation` packages. It also updates the versions of the `flutter_lints`, `riverpod_generator`, and `riverpod_lint` packages. This chore improves the project's dependencies and ensures compatibility with the latest versions of the packages.