chore: Update menu_cards.dart to fix color and height inconsistencies
This commit updates the `menu_cards.dart` file to fix color and height inconsistencies in the `_MenuCardsState` class. The color of the `moreText` widget is changed to `Colors.blue` and the height of the `SizedBox` widget is increased to `180`. This code improvement ensures a consistent visual presentation of the menu cards in the app.