feat: Add searchMealByName method to ServiceApi
This commit adds the `searchMealByName` method to the `ServiceApi` class. This method allows users to search for meals by name by making a GET request to the specified URL with the meal name as a parameter. The response is then parsed and returned as a list of `DetailMealModel` objects. This feature enhancement improves the functionality of the app by providing a convenient way to search for specific meals.